% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulindose.R
\name{insulindose}
\alias{insulindose}
\title{calculate insulin dose with default sensitivity factor and carb ratio}
\usage{
insulindose(glucose, carb)
}
\arguments{
\item{glucose}{current blood glucose reading (mg/dL)}

\item{carb}{carbohydrate intake (gram)}
}
\value{
insulin dose (unit)
}
\description{
calculate insulin base based on blood glucose and carb intake.
}
\details{
This function predicts the correct insulin dose by taking into account current glucose level and carbohydrate intake.
Insulin dose is equal to (current glucose level-target glucose level)/sensitivity factor+carbohydrate intake/carb ratio.
If the glucose level is less than 80, it will show the warning message.
Conventionally, target glucose is 120, sensitivity factor is 50 and carb ratio is 12.
}
\examples{
insulindose(170,48)
}
\author{
Chi Tang
}
