% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulindose2.R
\name{insulindose2}
\alias{insulindose2}
\title{calculate insulin dose with unknown sensitivity factor and carb ratio}
\usage{
insulindose2(glucose, carb, insulin)
}
\arguments{
\item{glucose}{current blood glucose reading (mg/dL)}

\item{carb}{carbohydrate intake (gram)}

\item{insulin}{total daily insulin requirement (unit)}
}
\value{
insulin dose (unit)
}
\description{
calculate insulin base based on blood glucose and carb intake when sensivitiy factor and carb ratio are not available.
}
\details{
This function predicts the correct insulin dose by taking into account current glucose level and carbohydrate intake.
Insulin dose is equal to (current glucose level-target glucose level)/sensitivity factor+carbohydrate intake/carb ratio.
If the glucose level is less than 80, it will show the warning message.
Conventionally, target glucose is 120
Since sensitivity factor and carb ratio are not available in this situation, we will need to estimate them based on total daily insulin requirement.
The formula is:
senitivity factor=2000/total daily insulin requirement
carb ratio=450/total daily insulin requirement
}
\examples{
insulindose(170,48,40)
}
\author{
Chi Tang
}
